Clean up admin native survey settings
